refacto upload methods
