+++ src/services/FileService.js
@@ -1,6 +1,15 @@
+import async from "async";
+
 import { FILE_TYPE } from "../config/files.js";
 import { download } from "../utils/download.js";
 import { segregate } from "../utils/file-structure.js";
+import {
+  getPaths,
+  handleInputFiles,
+  matchFoldersAndDocs,
+  createTreeFromPaths,
+  handleDragAndDropFile
+} from "../utils/files.js";
 
 import apiClient from "./api-client.js";
 import { ERRORS, RuntimeError, ErrorService } from "./ErrorService.js";
@@ -211,6 +220,43 @@ class FileService {
     } catch (error) {
       throw new RuntimeError(ERRORS.FILE_VERSIONS_MAKE_HEAD_ERROR, error);
     }
+  }
+
+  async createFolderStructure(project, currentFolder, files) {
+    let currentFiles = [];
+
+    if (files.isDirectory) {
+      currentFiles.push(await handleDragAndDropFile(files));
+    } else {
+      currentFiles.push(handleInputFiles(files));
+    }
+
+    return async.map(currentFiles, async folder => {
+      const paths = getPaths(folder);
+      const tree = createTreeFromPaths(currentFolder, paths);
+
+      let DMSTree;
+
+      try {
+        DMSTree = await apiClient.collaborationApi.createDMSTree(
+          project.cloud.id,
+          project.id,
+          [tree]
+        );
+      } catch (error) {
+        ErrorService.handleError(
+          new RuntimeError(ERRORS.FILE_STRUCTURE_CREATE_ERROR, error)
+        );
+        return {};
+      }
+
+      return {
+        type: FILE_TYPE.FOLDER,
+        name: DMSTree[0].name,
+        size: folder.reduce((a, b) => a + b.file?.size ?? 0, 0),
+        files: matchFoldersAndDocs(DMSTree, folder)
+      };
+    });
   }
 }
